* Fix build errors related to implicit function declarations.

+++ apps/VBuilder/src/w_main.c
@@ -40,7 +40,7 @@ void VbSetEditedControl(DesignerControl* pControl)
 #define SELECT_CTL_WIDTH  (150)
 #define SELECT_CTL_HEIGHT (200)
 
-void CALLBACK PrgVbSelectCtlProc(Window* pWindow, int messageType, int parm1, int parm2)
+void CALLBACK PrgVbSelectCtlProc(Window* pWindow, int messageType, long parm1, long parm2)
 {
 	switch (messageType)
 	{
